requirements.txt
flask==2.0.1
python-dotenv==0.19.0
schedule==1.1.0
newsapi-python==0.2.6
notion-client==1.0.0
requests==2.26.0
openai==0.27.0
beautifulsoup4==4.10.0
werkzeug==2.0.1
